- Implement CLI commands - Implement campaign levels from a game_data / patch_data Sound Shapes folder - Refactor community level import - Fix bug where Level Analysis would fail making it impossible to upload a level under certain circumstances - Refactor level publishing - Replace unix dates in API requests and responses with DateTimeOffsets
Fixes #200

+++ SoundShapesServer/Types/Levels/SSLevel/SSLevelEntity.cs
@@ -8,10 +8,10 @@ public class SSLevelEntity
 {
     [JsonProperty("entType")] public string EntityType { get; set; } = null!;
     public int Uid { get; set; }
-    public int MinScreenX { get; set; }
-    public int MinScreenY { get; set; }
-    public int MaxScreenX { get; set; }
-    public int MaxScreenY { get; set; }
+    public double MinScreenX { get; set; }
+    public double MinScreenY { get; set; }
+    public double MaxScreenX { get; set; }
+    public double MaxScreenY { get; set; }
     public float PosX { get; set; }
     public float PosY { get; set; }
